requirements.txt
absl-py==2.3.1
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.75.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.20
torch==2.7.1+cu118
torchinfo==1.8.0
torchvision==0.22.1+cu118
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
